Update network-stp.rst
Switch 32 has BDPU : [R=9, C=6, T=32.?] which is better than [R=9,C=7,T=19.2] because of the cost 6<7 (cost = 7 because 6 in the received BPDU but +1 for the link). 
For [R=9,C=12,T=42.1], cost : 6 < 12 thus the port towards switch 42 is designated. 
For [R=9,C=5,T=46.3], costs are equal with switch 32 BPDU cost but, T=32 < T=46 so the port towards 46 si also designated. 
Thus the answer should be all the port are designated.