address correction
In question 6, when the destination is R1, R4 should send the packet to R3. However, in the "correct" answer, is sends the packet to its own interface : 2001:db8:1341:34::4. I've changed this address which is now 2001:db8:1341:34::3. It correspond to the East interface of R3 as in the situation where  R3 is the destination.